fix mpvpaper toggler

+++ scripts/toggle-mpv.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
 
-killall mpvpaper
-notify-send "Mpvpaper" "Toggling.."
+if pidof mpvpaper > /dev/null; then
+    notify-send "Mpvpaper" "Disable.."
+    killall mpvpaper
+else
+    notify-send "Mpvpaper" "Enable.."
+    /home/nkee/.config/kr0ts/scripts/mpvwalls-daemon.sh &
+fi
